requirements.txt
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
pytorch==2.0.1
mlflow==2.4.1
fastapi==0.95.2
pytest==7.3.1
black==23.3.0
flake8==6.0.0
python-dotenv==1.0.0
hydra-core==1.3.2
wandb==0.15.4
dvc==2.58.2
great-expectations==0.16.13
prometheus-client==0.17.0
kubernetes==26.1.0 